Move Mutation and MutationSet to interfaces.py
