+++ api/ai/priority_algorithm/mutations/local_max_random.py
@@ -2,7 +2,7 @@ import random
 from typing import List, Dict
 
 from api.ai.priority_algorithm.custom_dataclasses import PriorityTeamSet
-from api.ai.priority_algorithm.mutations.mutation import Mutation
+from api.ai.priority_algorithm.mutations.interfaces import Mutation
 from api.ai.priority_algorithm.mutations.utils import (
     get_available_priority_teams,
     score,
