+++ api/ai/priority_algorithm/mutations/random_slice.py
@@ -3,7 +3,7 @@ from random import shuffle
 from typing import List, Dict
 
 from api.ai.priority_algorithm.custom_dataclasses import PriorityTeamSet
-from api.ai.priority_algorithm.mutations.mutation import Mutation
+from api.ai.priority_algorithm.mutations.interfaces import Mutation
 from api.ai.priority_algorithm.mutations.utils import get_available_priority_teams
 from api.ai.priority_algorithm.priority.interfaces import Priority
 from api.dataclasses.student import Student
