+++ benchmarking/runs/concentrate_many_attributes.py
@@ -10,7 +10,7 @@ from api.ai.interfaces.algorithm_config import (
     PriorityAlgorithmConfig,
 )
 from api.ai.priority_algorithm.mutations.local_max import LocalMaxMutation
-from api.ai.priority_algorithm.mutations.mutation_set import MutationSet
+from api.ai.priority_algorithm.mutations.interfaces import MutationSet
 from api.ai.priority_algorithm.mutations.random_swap import RandomSwapMutation
 from benchmarking.data.simulated_data.mock_student_provider import (
     MockStudentProvider,
