+++ tests/test_api/test_ai/test_priority_algorithm/test_mutations/test_mutations.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 
 from api.ai.priority_algorithm.custom_dataclasses import PriorityTeamSet, PriorityTeam
 from api.ai.priority_algorithm.mutations.local_max import LocalMaxMutation
-from api.ai.priority_algorithm.mutations.mutation_set import MutationSet
+from api.ai.priority_algorithm.mutations.interfaces import MutationSet
 from api.ai.priority_algorithm.mutations.random_swap import RandomSwapMutation
 from api.ai.priority_algorithm.priority_algorithm import (
     create_student_dict,
